Kiro/proofOfFace/ai-service/requirements.txt
# Flask Web Framework
Flask==3.0.0
flask-cors==4.0.0

# Face Recognition and Computer Vision
face-recognition==1.3.0
numpy==1.24.0
Pillow==10.0.0
opencv-python==4.8.1.78

# Cryptography and Security
cryptography==41.0.0

# HTTP and API
requests==2.31.0

# Environment and Configuration
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0